app/backend/requirements.txt
transformers
torch
fastapi
Pillow
numpy
uvicorn
diffusers
accelerate>=0.26.0
Pillow

